search bar updates
