+++ static/css/base.scss
@@ -21,6 +21,9 @@ $full-footer-bg: #fafcff;
 
 :root {
   --font-family: "Roboto", "Helvetica Neue", Arial, sans-serif !important;
+  --font-family-sans-serif: "Google Sans Text", -apple-system, BlinkMacSystemFont,
+  "Segoe UI", Roboto, "Helvetica Neue", Arial, sans-serif, "Apple Color Emoji",
+  "Segoe UI Emoji", "Segoe UI Symbol"
   --breakpoint-xl: 1040px;
 }
